+++ src/extension/Cycode.VisualStudio.Extension.Shared/Cli/DTO/CliIgnoreType.cs
@@ -3,5 +3,6 @@
 public enum CliIgnoreType {
     Value,
     Path,
-    Rule
+    Rule,
+    Cve,
 }