+++ src/extension/Cycode.VisualStudio.Extension.Shared/Components/ViolationCards/ScaViolationCardControl.xaml.cs
@@ -1,6 +1,9 @@
+using System.Windows;
+using Cycode.VisualStudio.Extension.Shared.Cli.DTO;
 using Cycode.VisualStudio.Extension.Shared.Cli.DTO.ScanResult.Sca;
 using Cycode.VisualStudio.Extension.Shared.Helpers;
 using Cycode.VisualStudio.Extension.Shared.Icons;
+using Cycode.VisualStudio.Extension.Shared.Services;
 
 namespace Cycode.VisualStudio.Extension.Shared.Components.ViolationCards;
 
@@ -11,9 +14,16 @@ public partial class ScaViolationCardControl {
     private const int _summaryRowIndex = 8;
     private const int _customRemediationGuidelinesRowIndex = 9;
     private const int _cycodeRemediationGuidelinesRowIndex = 10;
+    private const int _actionsButtonHrRowIndex = 11;
+    private const int _ignoreThisViolationButtonRowIndex = 12;
+    
+    private readonly ICycodeService _cycodeService = ServiceLocator.GetService<ICycodeService>();
+
+    private readonly ScaDetection _detection;
 
     public ScaViolationCardControl(ScaDetection detection) {
         InitializeComponent();
+        _detection = detection;
 
         Header.Icon.Source = ExtensionIcons.GetCardSeverityBitmapSource(detection.Severity);
         Header.Title.Text = detection.DetectionDetails.Alert?.Summary ?? detection.Message;
@@ -64,5 +74,16 @@ public partial class ScaViolationCardControl {
             CycodeGuidelines.Markdown = detection.DetectionDetails.RemediationGuidelines;
             GridHelper.ShowRow(Grid, _cycodeRemediationGuidelinesRowIndex);
         }
+        
+        if (string.IsNullOrEmpty(detection.DetectionDetails.Alert?.CveIdentifier)) {
+            GridHelper.HideRow(Grid, _actionsButtonHrRowIndex);
+            GridHelper.HideRow(Grid, _ignoreThisViolationButtonRowIndex);
+        }
+    }
+    
+    private async void IgnoreButton_OnClickAsync(object sender, RoutedEventArgs e) {
+        await _cycodeService.ApplyDetectionIgnoreAsync(
+            CliScanType.Sca, CliIgnoreType.Cve, _detection.DetectionDetails.Alert?.CveIdentifier
+        );
     }
 }